requirements.txt
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
PyYAML==6.0.2
vaderSentiment==3.3.2
fastapi==0.115.12
uvicorn==0.34.0
streamlit==1.44.1